feat: Extract from switchyard-server into switchyard-runner
These pieces will be shared by integrations, particularly the NeMo-Relay
plugin.

Implements: https://github.com/NVIDIA-NeMo/Switchyard/issues/514
