+++ crates/switchyard-runner/Cargo.toml
@@ -1,0 +1,30 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "switchyard-runner"
+version.workspace = true
+description = "Configured routing execution for Switchyard serving surfaces"
+authors.workspace = true
+edition.workspace = true
+homepage = "https://github.com/NVIDIA-NeMo/Switchyard"
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+publish = ["crates-io"]
+
+[dependencies]
+libsy = { package = "switchyard-libsy", path = "../libsy", version = "0.2.0" }
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+switchyard-llm-client.workspace = true
+switchyard-protocol.workspace = true
+thiserror.workspace = true
+toml = "1.1"
+tracing.workspace = true
+
+[dev-dependencies]
+async-trait.workspace = true
+futures-util.workspace = true
+tokio.workspace = true
