+++ crates/switchyard-runner/tests/route.rs
@@ -1,0 +1,164 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use std::collections::BTreeMap;
+use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::{Arc, Mutex};
+
+use async_trait::async_trait;
+use futures_util::StreamExt;
+use switchyard_llm_client::{ClientRouter, RunObservation};
+use switchyard_protocol::{
+    LlmClientError, LlmResponse, ModelId, Request, Response, RoutedLlmClient, text_request,
+    text_response,
+};
+use switchyard_runner::{AlgorithmSpec, ModelCapabilities, Route};
+
+struct StubClient;
+
+#[async_trait]
+impl RoutedLlmClient for StubClient {
+    async fn call(&self, request: Request) -> Result<Response, LlmClientError> {
+        Ok(Response {
+            llm_response: LlmResponse::Agg(text_response(
+                request.llm_request.model.clone(),
+                "plugin response",
+            )),
+            metadata: None,
+        })
+    }
+}
+
+fn plugin_route(client: Arc<dyn RoutedLlmClient>) -> Route {
+    let spec = AlgorithmSpec::Passthrough {
+        target: "semantic-target".to_string(),
+        subagents: None,
+    };
+    let targets = BTreeMap::from([(
+        "semantic-target".to_string(),
+        ModelId::from("semantic-target"),
+    )]);
+    let algorithm = spec
+        .build("switchyard", &targets)
+        .expect("identity target map should build");
+    let clients = ClientRouter::new(
+        BTreeMap::from([(ModelId::from("semantic-target"), client)])
+            .into_iter()
+            .collect(),
+    );
+    Route::new(
+        algorithm,
+        clients,
+        None,
+        ModelCapabilities::default(),
+        None,
+        Vec::new(),
+    )
+}
+
+#[tokio::test]
+async fn plugin_shaped_route_executes_without_runner_model_or_toml() {
+    let route = plugin_route(Arc::new(StubClient));
+    let observations = Arc::new(Mutex::new(Vec::new()));
+    let observer = {
+        let observations = Arc::clone(&observations);
+        Arc::new(move |observation| observations.lock().unwrap().push(observation))
+    };
+    let request = Request {
+        llm_request: text_request(Some("arbitrary-upstream-model".to_string()), "hello"),
+        ..Request::default()
+    };
+
+    let output = route
+        .execute(request, Some(observer))
+        .await
+        .expect("route should execute");
+
+    assert_eq!(output.selected_model, "semantic-target");
+    assert_eq!(
+        output
+            .response
+            .llm_response
+            .as_agg()
+            .unwrap()
+            .model
+            .as_deref(),
+        Some("semantic-target")
+    );
+    assert!(
+        observations
+            .lock()
+            .unwrap()
+            .iter()
+            .any(|observation| { matches!(observation, RunObservation::AnswerCall(_)) })
+    );
+    assert!(
+        observations
+            .lock()
+            .unwrap()
+            .iter()
+            .any(|observation| { matches!(observation, RunObservation::RoutingOverhead(_)) })
+    );
+}
+
+struct LazyStreamClient {
+    polls: Arc<AtomicUsize>,
+}
+
+#[async_trait]
+impl RoutedLlmClient for LazyStreamClient {
+    async fn call(&self, _request: Request) -> Result<Response, LlmClientError> {
+        let polls = Arc::clone(&self.polls);
+        let stream = futures_util::stream::poll_fn(move |_| {
+            polls.fetch_add(1, Ordering::SeqCst);
+            std::task::Poll::Ready(None)
+        })
+        .boxed();
+        Ok(Response {
+            llm_response: LlmResponse::Stream(stream),
+            metadata: None,
+        })
+    }
+}
+
+#[tokio::test]
+async fn route_returns_stream_without_polling_it() {
+    let polls = Arc::new(AtomicUsize::new(0));
+    let route = plugin_route(Arc::new(LazyStreamClient {
+        polls: Arc::clone(&polls),
+    }));
+    let request = Request {
+        llm_request: text_request(None, "hello"),
+        ..Request::default()
+    };
+
+    let output = route
+        .execute(request, None)
+        .await
+        .expect("stream handle should be returned");
+
+    assert!(matches!(
+        output.response.llm_response,
+        LlmResponse::Stream(_)
+    ));
+    assert_eq!(polls.load(Ordering::SeqCst), 0);
+}
+
+#[test]
+fn algorithm_build_reports_unknown_configured_target() {
+    let spec = AlgorithmSpec::Random {
+        targets: vec!["missing".to_string()],
+        weights: None,
+        seed: None,
+    };
+
+    let error = match spec.build("plugin", &BTreeMap::new()) {
+        Ok(_) => panic!("unknown target should fail"),
+        Err(error) => error,
+    };
+
+    assert_eq!(
+        error.to_string(),
+        "route plugin references unknown target missing"
+    );
+}
