+++ crates/switchyard-server/Cargo.toml
@@ -32,11 +32,10 @@ opentelemetry-prometheus = "0.32"
 opentelemetry_sdk = { version = "0.32", default-features = false, features = ["metrics", "trace"] }
 parking_lot.workspace = true
 prometheus = "0.14"
-reqwest.workspace = true
 serde.workspace = true
-toml = "1.1"
 switchyard-llm-client.workspace = true
 switchyard-protocol.workspace = true
+switchyard-runner.workspace = true
 serde_json.workspace = true
 switchyard-translation.workspace = true
 humantime = "2.4"
