+++ crates/switchyard-server/src/lib.rs
@@ -32,19 +32,21 @@ use axum::response::{IntoResponse, Response};
 use axum::routing::{get, post};
 use axum::{Extension, Json, Router};
 use axum_server::tls_rustls::RustlsConfig;
-use libsy::{Algorithm, CallModel, LibsyError, RoutingOutcome, drive};
+use libsy::{Algorithm, LibsyError, RoutingOutcome};
 use parking_lot::Mutex;
 use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
-use switchyard_llm_client::{ClientRouter, RunObservation, RunObserver, TranslatingLlmClient};
+use switchyard_llm_client::{ClientRouter, RunObservation, RunObserver};
 use switchyard_protocol::{LlmClientError, Metadata, ModelId, Request, Usage};
+use switchyard_runner::{
+    CallerAuthKind, DecisionTarget, ModelCapabilities, Route, RunOutput, Runner, RunnerError,
+};
 use tokio::net::{TcpListener, TcpSocket};
 use tokio::task;
 use tracing::{Instrument, Level};
 
 use switchyard_translation::{WireFormat, decode_request, encode_aggregated_response};
 
-use crate::config::ServerConfig;
 use crate::response::into_http_response;
 use crate::stats::{StatsAccumulator, StatsSnapshot, prefix_probe, tracking_enabled_from_env};
 
@@ -86,106 +88,44 @@ impl Display for ServerError {
 
 impl Error for ServerError {}
 
+impl From<RunnerError> for ServerError {
+    fn from(error: RunnerError) -> Self {
+        Self::new(error.to_string())
+    }
+}
+
 /// Result returned by server setup and lifecycle operations.
 pub type ServerResult<T> = std::result::Result<T, ServerError>;
 
-/// Capabilities that one route advertises on `GET /v1/models`.
-///
-/// An unset capability is undeclared: it serializes as `null` in the OpenAI
-/// `data` entry, and the Codex entry falls back to a safe default for it.
-#[derive(Clone, Copy, Default)]
-struct ModelCapabilities {
-    context_window: Option<u32>,
-    tool_calling: Option<bool>,
-    // Whether the routed model takes reasoning controls. The server cannot probe
-    // this, so a route opts in via config; undeclared routes advertise as
-    // non-reasoning to Codex (fail closed).
-    reasoning: Option<bool>,
-}
-
-/// A registered algorithm route and its server-owned endpoint metadata.
-struct RouteEntry {
-    algorithm: Arc<dyn Algorithm>,
-    /// Resolves each offloaded call to the client configured for the target the algorithm
-    /// selected. A route is a synthetic model with no upstream of its own, so this is a
-    /// per-target lookup, never one client serving the whole route.
-    target_clients: ClientRouter,
-    caller_auth: Option<CallerAuthKind>,
-    capabilities: ModelCapabilities,
-    count_tokens_target: Option<CountTokensTarget>,
-    config_name: Option<String>,
-}
-
 /// Decision result using the same selected/fallback order as libsy.
 #[derive(Serialize)]
-struct DecisionResponse<'a> {
-    selected: DecisionTargetResponse<'a>,
-    fallbacks: Vec<DecisionTargetResponse<'a>>,
+struct DecisionResponse {
+    selected: DecisionTargetResponse,
+    fallbacks: Vec<DecisionTargetResponse>,
     #[serde(skip_serializing_if = "Option::is_none")]
     response: Option<Value>,
 }
 
-/// Borrowed response view over one target's existing configuration.
+/// Response view over one target's existing configuration.
 #[derive(Serialize)]
-struct DecisionTargetResponse<'a> {
-    target: &'a str,
-    model: &'a ModelId,
-    llm_client: DecisionLlmClientResponse<'a>,
-    extra_body: &'a BTreeMap<String, Value>,
+struct DecisionTargetResponse {
+    target: String,
+    model: ModelId,
+    llm_client: DecisionLlmClientResponse,
+    extra_body: BTreeMap<String, Value>,
 }
 
 /// Non-secret client settings needed to call a selected model.
 #[derive(Serialize)]
-struct DecisionLlmClientResponse<'a> {
+struct DecisionLlmClientResponse {
     format: WireFormat,
-    base_url: &'a str,
-}
-
-/// Caller credential family required by forwarded-auth backends.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum CallerAuthKind {
-    Anthropic,
-    OpenAi,
-}
-
-impl CallerAuthKind {
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::Anthropic => "anthropic",
-            Self::OpenAi => "openai",
-        }
-    }
-
-    const fn accepts(self, wire_format: WireFormat) -> bool {
-        matches!(
-            (self, wire_format),
-            (Self::Anthropic, WireFormat::AnthropicMessages)
-                | (
-                    Self::OpenAi,
-                    WireFormat::OpenAiChat | WireFormat::OpenAiResponses
-                )
-        )
-    }
-}
-
-/// Exact upstream model used by the server's Anthropic token-count endpoint.
-#[derive(Clone)]
-struct CountTokensTarget {
-    model: ModelId,
-    client: Arc<TranslatingLlmClient>,
-}
-
-impl CountTokensTarget {
-    async fn count_tokens(&self, request: Request) -> Result<Value, LlmClientError> {
-        self.client.count_tokens(&self.model, request).await
-    }
+    base_url: String,
 }
 
 /// Shared server state used by all endpoint handlers.
 #[derive(Clone)]
 pub struct ServerState {
-    routes: Arc<BTreeMap<ModelId, RouteEntry>>,
-    config: Option<Arc<ServerConfig>>,
+    runner: Arc<Runner>,
     metrics: prometheus::Registry,
     stats: StatsAccumulator,
     routing_log: Option<SharedRoutingLog>,
@@ -227,86 +167,42 @@ impl SharedRoutingLog {
 }
 
 impl ServerState {
-    /// Creates server state from route model IDs, their libsy algorithms, and the
-    /// per-target client routing each route's calls are resolved through.
-    pub fn new(
-        routes: impl IntoIterator<Item = (ModelId, Arc<dyn Algorithm>, ClientRouter)>,
-    ) -> ServerResult<Self> {
-        Self::new_with_capabilities(routes.into_iter().map(|(model, algorithm, clients)| {
-            (
-                model,
-                algorithm,
-                clients,
-                None,
-                ModelCapabilities::default(),
-                None,
-                None,
-            )
-        }))
+    /// Creates server state from route model IDs, algorithms, and per-target clients.
+    pub fn new(routes: Vec<(ModelId, Arc<dyn Algorithm>, ClientRouter)>) -> ServerResult<Self> {
+        let routes = routes
+            .into_iter()
+            .map(|(model, algorithm, clients)| {
+                (
+                    model,
+                    Route::new(
+                        algorithm,
+                        clients,
+                        None,
+                        ModelCapabilities::default(),
+                        None,
+                        Vec::new(),
+                    ),
+                )
+            })
+            .collect();
+        let runner = Runner::new(routes);
+        Self::from_runner(runner)
     }
 
-    fn new_with_capabilities(
-        routes: impl IntoIterator<
-            Item = (
-                ModelId,
-                Arc<dyn Algorithm>,
-                ClientRouter,
-                Option<CallerAuthKind>,
-                ModelCapabilities,
-                Option<CountTokensTarget>,
-                Option<String>,
-            ),
-        >,
-    ) -> ServerResult<Self> {
-        let mut entries = BTreeMap::new();
-        for (
-            model,
-            algorithm,
-            target_clients,
-            caller_auth,
-            capabilities,
-            count_tokens_target,
-            config_name,
-        ) in routes
-        {
-            let model = ModelId::from(model.trim());
-            if model.is_empty() {
-                return Err(ServerError::new("route model must not be empty"));
-            }
-            let entry = RouteEntry {
-                algorithm,
-                target_clients,
-                caller_auth,
-                capabilities,
-                count_tokens_target,
-                config_name,
-            };
-            if entries.insert(model.clone(), entry).is_some() {
-                return Err(ServerError::new(format!("duplicate route model {model}")));
-            }
-        }
-        if entries.is_empty() {
-            return Err(ServerError::new("at least one algorithm route is required"));
-        }
+    /// Creates HTTP-server state around an already configured runner.
+    pub fn from_runner(runner: Runner) -> ServerResult<Self> {
         let metrics = metrics::registry().map_err(ServerError::new)?;
         let stats = StatsAccumulator::new(
             metrics.clone(),
-            entries.values().map(|entry| entry.algorithm.name()),
+            runner.models().map(|model| model.algorithm),
         );
         Ok(Self {
-            routes: Arc::new(entries),
-            config: None,
+            runner: Arc::new(runner),
             metrics,
             stats,
             routing_log: None,
             track_cache_eligibility: tracking_enabled_from_env(),
         })
-    }
-
-    /// Retains the validated configuration used to describe decision results.
-    fn with_config(mut self, config: Arc<ServerConfig>) -> Self {
-        self.config = Some(config);
-        self
     }
 
     /// Enables durable per-request routing records at `path`.
@@ -317,57 +213,44 @@ impl ServerState {
 
     /// Returns the route model IDs served by the configured algorithms.
     pub fn models(&self) -> impl Iterator<Item = &str> {
-        self.routes.keys().map(ModelId::as_str)
+        self.runner.models().map(|model| model.id.as_str())
     }
 
     /// Returns the caller credential family used by `model`, if any.
-    pub fn caller_auth_kind(&self, model: &str) -> ServerResult<Option<&'static str>> {
-        self.route_for_model(model)
-            .map(|entry| entry.caller_auth.map(CallerAuthKind::as_str))
-            .ok_or_else(|| ServerError::new(format!("unknown route model {model:?}")))
+    pub fn caller_auth_kind(&self, model: &str) -> Option<&'static str> {
+        self.runner
+            .route(model)
+            .and_then(|route| route.caller_auth())
+            .map(|kind| kind.as_str())
     }
 
-    fn route_for_model(&self, model: &str) -> Option<&RouteEntry> {
-        self.routes.get(model)
+    fn route_for_model(&self, model: &str) -> Option<&Route> {
+        self.runner.route(model)
     }
 
-    /// Resolves the routing outcome through the target names configured for its route.
-    fn decision_response<'a>(
-        &'a self,
-        route: &'a RouteEntry,
+    fn decision_response(
+        &self,
+        route_model: &ModelId,
         outcome: &RoutingOutcome,
         response: Option<Value>,
-    ) -> Option<DecisionResponse<'a>> {
-        let config = self.config.as_ref()?;
-        let target_names = config.routing_target_names(route.config_name.as_ref()?)?;
-        let resolve = |model: &ModelId| {
-            let (target_name, target) = target_names
-                .iter()
-                .filter_map(|name| config.targets.get_key_value(*name))
-                .find(|(_, target)| target.id == *model)?;
-            let client = config.llm_clients.get(&target.llm_client)?;
-            Some(DecisionTargetResponse {
-                target: target_name,
-                model: &target.id,
-                llm_client: DecisionLlmClientResponse {
-                    format: client.format.wire_format(),
-                    base_url: client.base_url.as_str(),
-                },
-                extra_body: &target.extra_body,
-            })
+    ) -> Option<DecisionResponse> {
+        let description = self.runner.describe_decision(route_model, outcome)?;
+        let convert = |target: DecisionTarget| DecisionTargetResponse {
+            target: target.target,
+            model: target.model,
+            llm_client: DecisionLlmClientResponse {
+                format: target.format,
+                base_url: target.base_url,
+            },
+            extra_body: target.extra_body,
         };
         Some(DecisionResponse {
-            selected: resolve(&outcome.selected_model_id)?,
-            fallbacks: outcome
-                .fallback_models
-                .iter()
-                .map(resolve)
-                .collect::<Option<Vec<_>>>()?,
+            selected: convert(description.selected),
+            fallbacks: description.fallbacks.into_iter().map(convert).collect(),
             response,
         })
     }
 }
-
 /// Runtime options shared by server entry points.
 #[derive(Clone, Debug)]
 pub struct ServerRunOptions {
@@ -682,9 +565,15 @@ async fn decision(
         Ok(resolved) => resolved,
         Err(response) => return response,
     };
-    let mut outcome = match run_decision_only(route, request).await {
+    let route_model = request
+        .llm_request
+        .model
+        .as_deref()
+        .map(ModelId::from)
+        .unwrap_or_default();
+    let mut outcome = match route.decide(request).await {
         Ok(outcome) => outcome,
-        Err(error) => return algorithm_error(error),
+        Err(error) => return runner_error(error),
     };
     let response = match outcome.response.take() {
         Some(response) => {
@@ -706,33 +595,12 @@ async fn decision(
         }
         None => None,
     };
-    match state.decision_response(route, &outcome, response) {
+    match state.decision_response(&route_model, &outcome, response) {
         Some(response) => Json(response).into_response(),
         None => {
             server_error("routing outcome contains a model with no callable target configuration")
         }
     }
-}
-
-/// Completes routing-time calls and returns the outcome without serving its answer target.
-async fn run_decision_only(route: &RouteEntry, request: Request) -> libsy::Result<RoutingOutcome> {
-    drive(Arc::clone(&route.algorithm), request, |call| {
-        serve_decision_dependency(route, call)
-    })
-    .await
-}
-
-/// Serves a classifier or judge call that the decision depends on.
-async fn serve_decision_dependency(route: &RouteEntry, call: CallModel) -> libsy::Result<()> {
-    let model = call.models.first().cloned().ok_or(LibsyError::NoTargets)?;
-    let result = match route.target_clients.route(&model) {
-        Ok(client) => client
-            .call(call.request.clone())
-            .await
-            .map_err(|source| LibsyError::client_call(model, source)),
-        Err(source) => Err(LibsyError::client_call(model, source)),
-    };
-    call.respond(result)
 }
 
 /// Anthropic token counting against the route's explicitly configured target.
@@ -756,17 +624,16 @@ async fn anthropic_count_tokens(
         Ok(resolved) => resolved,
         Err(response) => return anthropic_error_response(response),
     };
-    let Some(target) = route.count_tokens_target.as_ref() else {
-        return anthropic_error_response(error_response(
+    anthropic_error_response(match route.count_tokens(request).await {
+        Ok(payload) => (StatusCode::OK, Json(payload)).into_response(),
+        Err(RunnerError::CountTokensUnsupported) => error_response(
             StatusCode::BAD_REQUEST,
             "route has no Anthropic target for token counting",
             "invalid_request_error",
             "count_tokens_unsupported",
-        ));
-    };
-    anthropic_error_response(match target.count_tokens(request).await {
-        Ok(payload) => (StatusCode::OK, Json(payload)).into_response(),
-        Err(error) => count_tokens_error(error),
+        ),
+        Err(RunnerError::Client(error)) => count_tokens_error(error),
+        Err(error) => server_error(error.to_string()),
     })
 }
 
@@ -867,7 +734,7 @@ fn resolve_route(
     metadata: Metadata,
     body: Value,
     wire_format: WireFormat,
-) -> std::result::Result<(&RouteEntry, Request), Response> {
+) -> std::result::Result<(&Route, Request), Response> {
     let llm_request = decode_request(wire_format, &body)
         .map_err(|error| invalid_body_error(StatusCode::BAD_REQUEST, error.to_string()))?;
     let requested_model = llm_request
@@ -890,8 +757,8 @@ fn resolve_route(
             "model_not_found",
         )
     })?;
-    if let Some(caller_auth) = route.caller_auth
-        && !caller_auth.accepts(wire_format)
+    if let Err(RunnerError::IncompatibleCallerFormat(caller_auth)) =
+        route.check_caller_format(wire_format)
     {
         let (provider, expected_endpoint) = match caller_auth {
             CallerAuthKind::Anthropic => ("Anthropic", "/v1/messages"),
@@ -929,17 +796,18 @@ async fn handle_llm_request(
     };
     // Only the Codex namespace mapping is needed downstream, not the whole request.
     let request_extensions = request.llm_request.extensions.clone();
-    let algorithm = Arc::clone(&route.algorithm);
-    let client_router = route.target_clients.clone();
     let observer = stats_observer(
         state.stats.clone(),
         state.routing_log.clone().zip(routing_log_context.clone()),
     );
-    let (selected_model, response) =
-        match switchyard_llm_client::run(algorithm, client_router, request, Some(observer)).await {
-            Ok(result) => result,
-            Err(error) => return algorithm_error(error),
-        };
+    let output = match route.execute(request, Some(observer)).await {
+        Ok(output) => output,
+        Err(error) => return runner_error(error),
+    };
+    let RunOutput {
+        selected_model,
+        response,
+    } = output;
     // The response carries the candidate that actually served it. Fall back to the routing
     // selection for algorithms that return a response without an offloaded model call.
     let served_model = response.served_model().cloned().or(Some(selected_model));
@@ -1069,6 +937,14 @@ fn algorithm_error(error: LibsyError) -> Response {
         return server_error(error.to_string());
     };
     client_error(source)
+}
+
+fn runner_error(error: RunnerError) -> Response {
+    match error {
+        RunnerError::Algorithm(error) => algorithm_error(error),
+        RunnerError::Client(error) => client_error(&error),
+        error => server_error(error.to_string()),
+    }
 }
 
 fn client_error(error: &LlmClientError) -> Response {
@@ -1235,9 +1111,9 @@ fn error_response(
 async fn models(State(state): State<ServerState>) -> Json<Value> {
     Json(model_list_payload(
         state
-            .routes
-            .iter()
-            .map(|(model, entry)| (model.as_str(), entry.capabilities)),
+            .runner
+            .models()
+            .map(|model| (model.id.as_str(), model.capabilities)),
     ))
 }
 
@@ -1322,7 +1198,8 @@ fn model_list_payload<'a>(
     entries: impl IntoIterator<Item = (&'a str, ModelCapabilities)>,
 ) -> Value {
     let entries = entries.into_iter().collect::<Vec<_>>();
-    let model_ids = entries.iter().map(|(model, _)| *model).collect::<Vec<_>>();
+    let mut model_ids = entries.iter().map(|(model, _)| *model).collect::<Vec<_>>();
+    model_ids.sort_unstable();
     let first_id = model_ids.first().copied();
     let last_id = model_ids.last().copied();
     json!({
